requirements.txt
Django>=4.2
django-jazzmin
djangorestframework-simplejwt
django-unfold>=0.18
gunicorn
Pillow
psycopg2-binary
djangorestframework
django-cors-headers
python-dotenv
channels
numpy
scikit-learn